Removing contacts and planned Migration
I will try to migrate this script from VBS to PowerShell 7. 
The original intent in keeping this code in VBS was to easily deploy it anywhere and get these data of interest.
I am deviating from that to make this script work for more recent version of Windows using the more relevant framework to retrieve system configuration.